Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int n;
-    scanf("%d",n)  
+    long long int n;
+    scanf("%lld",n)  
     int msb = (n >> 31) & 1;  
     if (msb == 1) {
         printf("Set");
